Support reading neutral mass from compound DB
Until now, El-MAVEN has been able to directly read adjusted m/z
value or a formula from which we calculate the neutral mass and
then adjust the m/z according to the session's polarity. Now a new
header, named "mass", can be provided in the compound database that
stores neutral mass of each compound which only needs to be
adjusted to get the m/z that needs to be searched for.