[Fix] Allow ignoring close event
When project saving feature was added, a prompt would appear asking
the user to save their session before the exit. Once the "close"
button of the main application was clicked, there was no way of
cancelling the exit event (with or without saving). With this
patch, the user will have the option to ignore the exit event by
either selecting "Cancel" button or pressing "Esc" when then save
prompt appears.