Upgrade to new releases using maintenancetool
With this commit, a mechanism has been set-up such that whenever
the GUI window starts up, the controller will check whether there
is a newer release available at El-MAVEN's update repository. The
user will be informed if the application can be updated and
prompted to begin the process. If they allow, the maintenancetool
executable will start and guide the user through the upgrade
process.